requirements.txt
APScheduler==3.6.3
asgiref==3.11.0
cachetools==4.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tornado==6.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
